split real and fake clocks
